+++ test/test_AD.jl
@@ -2,7 +2,7 @@ using KernelFunctions
 using Zygote, ForwardDiff
 using Test, LinearAlgebra
 using FiniteDifferences
-include("utils_AD.jl")
+
 dims = [10,5]
 
 A = rand(dims...)
